refactor sandbox command
