+++ cmd/root.go
@@ -25,6 +25,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/ks6088ts/template-go/cmd/sandbox"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -67,6 +68,9 @@ func init() {
 	// Cobra also supports local flags, which will only run
 	// when this action is called directly.
 	rootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+
+	// Register sub commands
+	registerSubCommands()
 }
 
 // initConfig reads in config file and ENV variables if set.
@@ -91,4 +95,10 @@ func initConfig() {
 	if err := viper.ReadInConfig(); err == nil {
 		fmt.Fprintln(os.Stderr, "Using config file:", viper.ConfigFileUsed())
 	}
+}
+
+// registerSubCommands registers sub commands
+func registerSubCommands() {
+	rootCmd.AddCommand(sandbox.GetCommand())
+	// Add other sub commands here
 }
