+++ cmd/sandbox/sandbox.go
@@ -19,7 +19,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-package cmd
+package sandbox
 
 import (
 	"fmt"
@@ -30,28 +30,13 @@ import (
 // sandboxCmd represents the sandbox command
 var sandboxCmd = &cobra.Command{
 	Use:   "sandbox",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "SandBox command",
+	Long:  `This is a sandbox command.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("sandbox called")
 	},
 }
 
-func init() {
-	rootCmd.AddCommand(sandboxCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// sandboxCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// sandboxCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+func GetCommand() *cobra.Command {
+	return sandboxCmd
 }
